+++ crates/backtest/src/lib.rs
@@ -67,5 +67,8 @@ pub mod exchange;
 pub mod execution_client;
 pub mod modules;
 
+#[cfg(feature = "streaming")]
+pub mod node;
+
 #[cfg(feature = "ffi")]
 pub mod ffi;